+++ packages/core/core-flows/src/auth/steps/index.ts
@@ -1,1 +1,2 @@
+export * from "./request-verification"
 export * from "./set-auth-app-metadata"